# docs/openapi/control-api-v1.yaml
openapi: 3.0.3
info:
  title: ProjectD Control API
  version: '1.0.0'
  description: |
    Hub Control API (`ac-data-backend`) for the central mod library, per-VPS local-cache
    availability, legacy disk inventory, and live lobby rosters.
    Auth: `X-Worker-Secret` (same as Convex worker secret / CONVEX_WORKER_SECRET).

    Central library: `GET /v1/mods`, `GET /v1/mods/{slug}`, `GET /v1/mods/availability`,
    `POST /v1/instances/{instanceId}/mods/ensure`.
    Legacy LOCAL CONTENT / INVENTORY: `GET /v1/instances/{instanceId}/mods/cars|tracks`
    (Redis disk scan — keep during Host cutover; do not use as the global library).

    Live write path: edge processes join/leave/server_status into shared Redis HUD keys
    (`ac:hud:presence:*`, `ac:hud:presence:roster:{instanceId}:{lobby}`). There is NO
    `POST /v1/agents/{instanceId}/live` — this API only reads Redis for live GETs.
    Mods write path: `POST /v1/agents/{instanceId}/mods` → Redis `instance:{id}:mods:*`.

    Path param `serverId` on live endpoints is the normalized AC lobby name, not a Convex
    `servers._id`. Prefer `?instanceId=` on multi-VPS.

    Server slots (Postgres `server_slots`): `GET/POST /v1/servers`, `POST /v1/servers/allocate`,
    apply-config / start / stop / restart by slot UUID. Requires DATABASE_URL.
    See docs/SERVER_PLATFORM.md and docs/CONTROL_API_V1.md.
servers:
  - url: https://hub.example.com
    description: Hub public base (CONTROL_API_URL / AC_DATA_BASE_URL)
security:
  - WorkerSecret: []
paths:
...
  /v1/agents/{instanceId}:
    get:
      summary: Debug — read last agent presence for instance
      operationId: getAgentPresence
      parameters:
        - $ref: '#/components/parameters/instanceId'
      responses:
        '200':
          description: Presence document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentPresenceGetResponse'
        '404':
          description: No presence
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/mods:
    get:
      summary: List central mod library (Postgres catalog)
      description: |
        Source of truth for mods/cars/tracks/versions/artifacts. Not VPS inventory.
        Each package may include imageUrl (one cover image) or null.
      operationId: listCentralMods
      parameters:
        - name: kind
          in: query
          schema:
            type: string
            enum: [car, track, weather, misc]
      responses:
        '200':
          description: Catalog packages with versions
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: DATABASE_URL not configured
  /v1/mods/availability:
    get:
      summary: Central library + local cache availability for a VPS
      operationId: getModsAvailability
      parameters:
        - name: instanceId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: |
            Each item has central=AVAILABLE and local=LOCAL|MISSING|DOWNLOADING|INSTALLING|ERROR.
            Match by artifactId + sha256 (exact version).
        '400':
          description: instanceId required
        '404':
          description: fleet edge not found for instanceId
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/mods/{slug}:
    get:
      summary: Get one central mod package by slug (or ac_content_slug)
      operationId: getCentralModBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Package + versions
        '404':
          description: Not found
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/servers:
    get:
      summary: List server slots (Postgres SoT)
      description: |
        Requires DATABASE_URL. Filter with `region`, `status` (idle|allocated|live|draining|error),
        `instanceId`. Used by Host to pick available capacity.
      operationId: listServerSlots
      parameters:
        - name: region
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [idle, allocated, live, draining, error]
        - name: instanceId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Slot list
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: DATABASE_URL not configured
    post:
      summary: Upsert / bootstrap a server slot
      operationId: upsertServerSlot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertServerSlotBody'
      responses:
        '200':
          description: Upserted
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/servers/allocate:
    post:
      summary: Allocate an idle slot in a region
      operationId: allocateServerSlot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [region]
              properties:
                region:
                  type: string
                presetRef:
                  type: string
                  description: Optional Convex preset id
      responses:
        '200':
          description: Allocated slot
        '409':
          description: No idle slot in region
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    WorkerSecret:
      type: apiKey
      in: header
      name: X-Worker-Secret
  parameters:
    instanceId:
      name: instanceId
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Missing or invalid X-Worker-Secret
  schemas:
...
    AgentPresenceGetResponse:
      type: object
      properties:
        ok:
          type: boolean
        instanceId:
          type: string
        presence:
          type: object
          additionalProperties: true
...
    UpsertServerSlotBody:
      type: object
      required: [instanceId, region, lobbyName, folderSlug]
      properties:
        instanceId:
          type: string
        region:
          type: string
        lobbyName:
          type: string
          description: AC lobby display / live key name
        folderSlug:
          type: string
          description: Edge server folder name (e.g. server-1)
        status:
          type: string
          enum: [idle, allocated, live, draining, error]